refactor(utils): renamed "downloadBlob" to "clientFileDownload"
