+++ webapp/src/routes/[[lang]]/my/organizations/[id]/credential-issuances/[serviceId]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { request } from '@api/downloadCredentialIssuer';
-	import { downloadBlob } from '$lib/utils/downloadBlob';
+	import { downloadBlob } from '$lib/utils/clientFileDownload';
 	import CopyImageButton from '$lib/components/copyImageButton.svelte';
 
 	import ServiceForm from '../_partials/serviceForm.svelte';
@@ -88,7 +88,13 @@
 		{:then qrimg}
 			<div class="flex gap-4 items-start">
 				<img src={qrimg} alt={m.Service_Qr_Code()} class="border rounded-lg" />
-				<CopyImageButton imageSrc={qrimg}>Copy QR code</CopyImageButton>
+				<div class="flex flex-col gap-2">
+					<CopyImageButton imageSrc={qrimg}>Copy QR code</CopyImageButton>
+					<Button color="alternative" size="sm">
+						<ArrowDownTray />
+						<span class="ml-2">Download QR code</span>
+					</Button>
+				</div>
 			</div>
 		{:catch error}
 			<p class="text-red-500">{JSON.stringify(error)}</p>
